Move the contents of tracker.js into app.js and remove tracker.js
